Refactor Discord bot into single output-only module
