--- rpc/discord/bsky/__init__.py
@@ -1,5 +1,0 @@
-from .services import discord_bsky_post_message_v1
-
-DISPATCHERS: dict[tuple[str, str], callable] = {
-  ("post", "1"): discord_bsky_post_message_v1,
-}
