+++ server/modules/discord_bot_module.py
@@ -1,9 +1,12 @@
 """Discord bot coordination module."""
 
-import logging, discord, asyncio, os
-from typing import IO, TYPE_CHECKING, Any
-from fastapi import FastAPI
+import asyncio, logging, os, time, discord
+from collections.abc import Awaitable, Callable
+from contextlib import suppress
+from typing import IO, TYPE_CHECKING, Any, NamedTuple
+
 from discord.ext import commands
+from fastapi import FastAPI
 
 try:  # pragma: no cover - platform dependent import
   import fcntl
@@ -15,32 +18,37 @@ try:  # pragma: no cover - platform dependent import
 except ImportError:  # pragma: no cover - non-Windows platforms
   msvcrt = None
 
-from . import BaseModule
-from .env_module import EnvModule
-from .db_module import DbModule
-from queryregistry.system.config.models import ConfigKeyParams
-from queryregistry.system.config import get_config_request
 from queryregistry.discord.guilds import (
   get_guild_request,
   list_guilds_request,
-  upsert_guild_request,
   update_credits_request,
+  upsert_guild_request,
 )
 from queryregistry.discord.guilds.models import (
   GuildIdParams,
   ListGuildsParams,
-  UpsertGuildParams,
   UpdateGuildCreditsParams,
+  UpsertGuildParams,
 )
-
+from queryregistry.system.config import get_config_request
+from queryregistry.system.config.models import ConfigKeyParams
 from server.helpers.logging import configure_discord_logging, remove_discord_logging, update_logging_level
 from server.routers.discord_events import register_discord_event_handlers
 
+from . import BaseModule
+from .db_module import DbModule
+from .env_module import EnvModule
+
 if TYPE_CHECKING:  # pragma: no cover
-  from .discord_output_module import DiscordOutputModule
   DiscordAuthModule = Any
-  SocialInputModule = Any
-  DiscordInputProvider = Any
+
+_SendCallable = Callable[[str], Awaitable[None]]
+
+
+class _QueuePayload(NamedTuple):
+  kind: str
+  target_id: int
+  message: str
 
 
 class DiscordBotModule(BaseModule):
@@ -51,48 +59,40 @@ class DiscordBotModule(BaseModule):
     self.bot: commands.Bot | None = None
     self.db: DbModule | None = None
     self.env: EnvModule | None = None
-    self.discord_output: "DiscordOutputModule" | None = None
     self.discord_auth: "DiscordAuthModule" | None = None
     self._task: asyncio.Task | None = None
-    self._user_counts: dict[int, int] = {}
-    self._guild_counts: dict[int, int] = {}
-    self.USER_RATE_LIMIT = 100
-    self.GUILD_RATE_LIMIT = 1000
+    self._worker_task: asyncio.Task[None] | None = None
+    self._message_size_limit = 1998
+    self._trickle_delay = 1.0
+    self._send_lock = asyncio.Lock()
+    self._outbound_queue: asyncio.Queue[_QueuePayload] = asyncio.Queue()
+    self._stats_lock = asyncio.Lock()
+    self._channel_stats: dict[int, dict[str, float | int]] = {}
+    self._user_stats: dict[int, dict[str, float | int]] = {}
+    self._aggregate_stats: dict[str, float | int] = {"messages": 0, "characters": 0, "last_sent_at": 0.0}
     self.owns_bot: bool = False
     self._lock_handle: IO[str] | None = None
     self.auth_module: "DiscordAuthModule" | None = None
-    self.output_module: "DiscordOutputModule" | None = None
-    self.social_input_module: "SocialInputModule" | None = None
-    self.input_provider: "DiscordInputProvider" | None = None
     if not getattr(self.app.state, "discord_bot", None):
       setattr(self.app.state, "discord_bot", self)
 
   async def startup(self):
-    self.env: EnvModule = self.app.state.env
+    self.env = self.app.state.env
     await self.env.on_ready()
-    self.db: DbModule = self.app.state.db
+    self.db = self.app.state.db
     await self.db.on_ready()
-    self.discord_output = getattr(self.app.state, "discord_output", None)
     self.discord_auth = getattr(self.app.state, "discord_auth", None) or getattr(self.app.state, "auth", None)
-    self.social_input_module = getattr(self.app.state, "social_input", None)
     setattr(self.app.state, "discord", self)
     if not self._acquire_bot_lock():
       logging.info("[DiscordBotModule] startup skipped: Discord bot already owned by another worker")
       self.mark_ready()
       return
+
     try:
       self.owns_bot = True
-      setattr(self.app.state, "discord_bot", self)
       self.secret = self.env.get("DISCORD_SECRET")
       self.bot = self._init_discord_bot('!')
       self.bot.app = self.app
-      register_discord_event_handlers(self)
-      update_logging_level(self.db.logging_level)
-      configure_discord_logging(self)
-      res = await self.db.run(get_config_request(ConfigKeyParams(key="DiscordSyschan")))
-      if not res.rows:
-        raise ValueError("Missing config value for key: DiscordSyschan")
-      self.syschan = int(res.rows[0]["element_value"] or 0)
       try:
         await self.bot.login(self.secret)
         self._task = asyncio.create_task(self.bot.connect())
@@ -102,7 +102,24 @@ class DiscordBotModule(BaseModule):
         if self._task:
           self._task.cancel()
         raise
+
+      register_discord_event_handlers(self)
+      update_logging_level(self.db.logging_level)
+      configure_discord_logging(self)
+      res = await self.db.run(get_config_request(ConfigKeyParams(key="DiscordSyschan")))
+      if not res.rows:
+        raise ValueError("Missing config value for key: DiscordSyschan")
+      self.syschan = int(res.rows[0]["element_value"] or 0)
+      if not self._worker_task:
+        self._worker_task = asyncio.create_task(self._queue_worker(), name="discord-output-worker")
+      setattr(self.app.state, "discord_bot", self)
+      setattr(self.app.state, "discord_output", self)
     except Exception:
+      if self._worker_task:
+        self._worker_task.cancel()
+        with suppress(asyncio.CancelledError):
+          await self._worker_task
+      self._worker_task = None
       if self._task and not self._task.cancelled():
         self._task.cancel()
       self._task = None
@@ -112,11 +129,20 @@ class DiscordBotModule(BaseModule):
       self.owns_bot = False
       if getattr(self.app.state, "discord_bot", None) is self:
         self.app.state.discord_bot = None
+      if getattr(self.app.state, "discord_output", None) is self:
+        self.app.state.discord_output = None
       raise
+
     logging.debug("Discord bot module loaded")
     self.mark_ready()
 
   async def shutdown(self):
+    if self._worker_task:
+      self._worker_task.cancel()
+      with suppress(asyncio.CancelledError):
+        await self._worker_task
+      self._worker_task = None
+
     if self.bot:
       await self.bot.close()
       self.bot = None
@@ -126,18 +152,20 @@ class DiscordBotModule(BaseModule):
       except asyncio.CancelledError:
         pass
       self._task = None
+
     remove_discord_logging(self)
     self._release_bot_lock()
     self.owns_bot = False
     if getattr(self.app.state, "discord_bot", None) is self:
       self.app.state.discord_bot = None
+    if getattr(self.app.state, "discord_output", None) is self:
+      self.app.state.discord_output = None
 
   def _init_discord_bot(self, prefix: str) -> commands.Bot:
     intents = discord.Intents.default()
     intents.guild_messages = True
     intents.guilds = True
     intents.members = True
-    intents.message_content = True
     return commands.Bot(command_prefix=prefix, intents=intents)
 
   def _acquire_bot_lock(self) -> bool:
@@ -188,97 +216,206 @@ class DiscordBotModule(BaseModule):
   def register_auth_module(self, module: "DiscordAuthModule") -> None:
     self.auth_module = module
 
-  def register_output_module(self, module: "DiscordOutputModule") -> None:
-    self.output_module = module
+  def configure_message_window(self, max_message_size: int) -> None:
+    if max_message_size <= 0:
+      raise ValueError("max_message_size must be greater than zero")
+    self._message_size_limit = max_message_size
 
-  def register_social_input_module(self, module: "SocialInputModule") -> None:
-    self.social_input_module = module
-
-  def register_input_provider(self, provider: "DiscordInputProvider") -> None:
-    self.input_provider = provider
-
-  def _get_output_module(self) -> "DiscordOutputModule | None":
-    if self.output_module:
-      return self.output_module
-    output = getattr(self.app.state, "discord_output", None)
-    if output:
-      self.output_module = output
-    return output
-
-  def _require_output_module(self) -> "DiscordOutputModule":
-    output = self._get_output_module()
-    if not output:
-      raise RuntimeError("DiscordOutputModule is not available")
-    return output
+  def configure_trickle_rate(self, delay_seconds: float) -> None:
+    if delay_seconds < 0:
+      raise ValueError("delay_seconds cannot be negative")
+    self._trickle_delay = delay_seconds
 
   async def send_channel_message(self, channel_id: int, message: str) -> None:
-    if not message:
-      return
-    output = self._require_output_module()
-    await output.send_to_channel(channel_id, message)
+    await self.send_to_channel(channel_id, message)
 
   async def send_user_message(self, user_id: int, message: str) -> None:
+    await self.send_to_user(user_id, message)
+
+  async def send_to_channel(self, channel_id: int, message: str) -> None:
     if not message:
       return
-    output = self._require_output_module()
-    await output.send_to_user(user_id, message)
+    channel = await self._resolve_channel(channel_id)
+    await self._deliver_in_chunks(channel.send, message)
+    await self._record_channel_throughput(channel_id, message)
+
+  async def send_to_user(self, user_id: int, message: str) -> None:
+    if not message:
+      return
+    user = await self._resolve_user(user_id)
+    await self._deliver_in_chunks(user.send, message)
+    await self._record_user_throughput(user_id, message)
 
   async def queue_channel_message(self, channel_id: int, message: str) -> None:
     if not message:
       return
-    output = self._require_output_module()
-    await output.queue_channel_message(channel_id, message)
+    await self._outbound_queue.put(_QueuePayload("channel", channel_id, message))
 
   async def queue_user_message(self, user_id: int, message: str) -> None:
     if not message:
       return
-    output = self._require_output_module()
-    await output.queue_user_message(user_id, message)
+    await self._outbound_queue.put(_QueuePayload("user", user_id, message))
+
+  async def wait_for_drain(self) -> None:
+    await self._outbound_queue.join()
+
+  async def get_throughput_snapshot(self) -> dict[str, dict]:
+    async with self._stats_lock:
+      return {
+        "aggregate": dict(self._aggregate_stats),
+        "channels": {cid: dict(stats) for cid, stats in self._channel_stats.items()},
+        "users": {uid: dict(stats) for uid, stats in self._user_stats.items()},
+      }
+
+  async def _resolve_channel(self, channel_id: int) -> discord.abc.Messageable:
+    if not self.bot:
+      raise RuntimeError("Discord bot is not available")
+    channel = self.bot.get_channel(channel_id)
+    if channel is None:
+      try:
+        channel = await self.bot.fetch_channel(channel_id)
+      except discord.DiscordException:
+        logging.exception("[DiscordBotModule] failed to fetch channel", extra={"channel_id": channel_id})
+        raise
+    if not hasattr(channel, "send"):
+      raise RuntimeError(f"Channel {channel_id} is not send-capable")
+    return channel
+
+  async def _resolve_user(self, user_id: int) -> discord.abc.Messageable:
+    if not self.bot:
+      raise RuntimeError("Discord bot is not available")
+    user = self.bot.get_user(user_id)
+    if user is None:
+      try:
+        user = await self.bot.fetch_user(user_id)
+      except discord.DiscordException:
+        logging.exception("[DiscordBotModule] failed to fetch user", extra={"user_id": user_id})
+        raise
+    if not hasattr(user, "send"):
+      raise RuntimeError(f"User {user_id} cannot receive messages")
+    return user
+
+  async def _queue_worker(self) -> None:
+    while True:
+      try:
+        payload = await self._outbound_queue.get()
+      except asyncio.CancelledError:
+        raise
+      try:
+        if payload.kind == "channel":
+          await self.send_to_channel(payload.target_id, payload.message)
+        else:
+          await self.send_to_user(payload.target_id, payload.message)
+      except asyncio.CancelledError:
+        raise
+      except Exception:  # pragma: no cover - observability path
+        logging.exception(
+          "[DiscordBotModule] failed to dispatch message",
+          extra={"kind": payload.kind, "target": payload.target_id},
+        )
+      finally:
+        self._outbound_queue.task_done()
+
+  async def _deliver_in_chunks(self, sender: _SendCallable, text: str) -> None:
+    chunks = self._yield_chunks(text, self._message_size_limit)
+    async with self._send_lock:
+      for chunk in chunks:
+        await sender(chunk)
+        if self._trickle_delay > 0:
+          await asyncio.sleep(self._trickle_delay)
+
+  async def _record_channel_throughput(self, channel_id: int, message: str) -> None:
+    await self._record_stats(self._channel_stats, channel_id, message)
+
+  async def _record_user_throughput(self, user_id: int, message: str) -> None:
+    await self._record_stats(self._user_stats, user_id, message)
+
+  async def _record_stats(self, bucket: dict[int, dict[str, float | int]], identifier: int, message: str) -> None:
+    async with self._stats_lock:
+      stats = bucket.setdefault(identifier, {"messages": 0, "characters": 0, "last_sent_at": 0.0})
+      now = time.time()
+      stats["messages"] += 1
+      stats["characters"] += len(message)
+      stats["last_sent_at"] = now
+      self._aggregate_stats["messages"] += 1
+      self._aggregate_stats["characters"] += len(message)
+      self._aggregate_stats["last_sent_at"] = now
+
+  @staticmethod
+  def _wrap_line(line: str, max_message_size: int) -> list[str]:
+    remaining = line
+    parts: list[str] = []
+    while remaining:
+      if len(remaining) <= max_message_size:
+        parts.append(remaining)
+        break
+      window = remaining[:max_message_size]
+      split = window.rfind(" ")
+      if split <= 0:
+        split = max_message_size
+      parts.append(remaining[:split])
+      remaining = remaining[split:]
+    if not parts:
+      parts.append("")
+    return parts
+
+  @classmethod
+  def _yield_chunks(cls, text: str, max_message_size: int) -> list[str]:
+    normalized = text.replace("\r\n", "\n").replace("\r", "\n")
+    lines = normalized.split("\n")
+    chunks: list[str] = []
+    buffer = ""
+    for raw_line in lines:
+      line = raw_line
+      if not line.strip() and not buffer.strip():
+        continue
+      wrapped = cls._wrap_line(line, max_message_size)
+      for segment in wrapped[:-1]:
+        candidate = segment if not buffer else buffer + "\n" + segment
+        if candidate.strip() and candidate.strip() != "---":
+          chunks.append(candidate)
+        buffer = ""
+      last_segment = wrapped[-1]
+      candidate = last_segment if not buffer else buffer + ("\n" + last_segment if last_segment else "\n")
+      if len(candidate) <= max_message_size:
+        buffer = candidate
+      else:
+        if buffer.strip() and buffer.strip() != "---":
+          chunks.append(buffer)
+        buffer = last_segment
+    if buffer.strip() and buffer.strip() != "---":
+      chunks.append(buffer)
+    return chunks
 
   async def _try_send_channel(self, channel_id: int, message: str) -> bool:
     try:
-      await self.send_channel_message(channel_id, message)
+      await self.send_to_channel(channel_id, message)
       return True
     except Exception:
       logging.exception(
-        "[DiscordBotModule] failed to send via DiscordOutputModule",
+        "[DiscordBotModule] failed to send channel message",
         extra={"channel_id": channel_id},
       )
-    return False
+      return False
 
   async def _try_send_user(self, user_id: int, message: str) -> bool:
     try:
-      await self.send_user_message(user_id, message)
+      await self.send_to_user(user_id, message)
       return True
     except Exception:
       logging.exception(
-        "[DiscordBotModule] failed to send user message via DiscordOutputModule",
+        "[DiscordBotModule] failed to send user message",
         extra={"user_id": user_id},
       )
-    return False
-
-  def bump_rate_limits(self, guild_id: int, user_id: int):
-    u = self._user_counts.get(user_id, 0) + 1
-    g = self._guild_counts.get(guild_id, 0) + 1
-    self._user_counts[user_id] = u
-    self._guild_counts[guild_id] = g
-    if u >= int(self.USER_RATE_LIMIT * 0.8):
-      logging.info("[DiscordBot] user nearing rate limit", extra={"user_id": user_id, "count": u})
-    if g >= int(self.GUILD_RATE_LIMIT * 0.8):
-      logging.info("[DiscordBot] guild nearing rate limit", extra={"guild_id": guild_id, "count": g})
+      return False
 
   async def send_sys_message(self, message: str):
     if not self.syschan:
       return
-    if await self._try_send_channel(self.syschan, message):
-      return
-    if not self.bot:
-      return
-    channel = self.bot.get_channel(self.syschan)
-    if channel:
-      from server.helpers.logging import split_message
-      for part in split_message(message):
-        await channel.send(part)
+    try:
+      await self.send_to_channel(self.syschan, message)
+    except Exception:
+      logging.exception("[DiscordBotModule] failed to send sys message")
 
   def _map_guild_row(self, row: dict[str, Any], guild_id: str = "") -> dict[str, Any]:
     return {
@@ -343,6 +480,7 @@ class DiscordBotModule(BaseModule):
     guilds = await self.list_guild_records()
     logging.info("[DiscordBotModule] synced %d guilds", synced)
     return {"synced": synced, "guilds": guilds}
+
 
 class DiscordModule(DiscordBotModule):
   """Backward-compatible alias for the DiscordBotModule."""
